requirements.txt
streamlit==1.32.2
pandas==2.2.1
psycopg2-binary==2.9.9
plotly==5.19.0
python-dotenv==1.0.1
openpyxl==3.1.2
